Restore the original test_accel.py for backwards compatibility
It contains just the decorators for use with nosetests, and no actual
tests.
