+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 asyncio_mode = auto
+addopts = --ignore=src/katsdpsigproc/test  # It's utilities for downstream packages, not tests
